Refactor to standalone module; update spacing on invalid wf state editor view
