+++ public/pages/workflow_detail/resizable_workspace.tsx
@@ -286,27 +286,31 @@ export function ResizableWorkspace(props: ResizableWorkspaceProps) {
       )}
     </Formik>
   ) : (
-    <>
-      <EuiEmptyPrompt
-        iconType={'cross'}
-        title={<h2>Unable to view workflow details</h2>}
-        titleSize="s"
-        body={
-          <>
-            <EuiText>
-              Only valid workflows created from this OpenSearch Dashboards
-              application are editable and viewable.
-            </EuiText>
-          </>
-        }
-      />
-      <EuiCodeBlock language="json" fontSize="m" isCopyable={false}>
-        {JSON.stringify(
-          reduceToTemplate(props.workflow as Workflow),
-          undefined,
-          2
-        )}
-      </EuiCodeBlock>
-    </>
+    <EuiFlexGroup direction="column">
+      <EuiFlexItem grow={3}>
+        <EuiEmptyPrompt
+          iconType={'cross'}
+          title={<h2>Unable to view workflow details</h2>}
+          titleSize="s"
+          body={
+            <>
+              <EuiText>
+                Only valid workflows created from this OpenSearch Dashboards
+                application are editable and viewable.
+              </EuiText>
+            </>
+          }
+        />
+      </EuiFlexItem>
+      <EuiFlexItem grow={7}>
+        <EuiCodeBlock language="json" fontSize="m" isCopyable={false}>
+          {JSON.stringify(
+            reduceToTemplate(props.workflow as Workflow),
+            undefined,
+            2
+          )}
+        </EuiCodeBlock>
+      </EuiFlexItem>
+    </EuiFlexGroup>
   );
 }
