+++ public/pages/workflows/import_workflow/index.ts
@@ -1,0 +1,6 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+export { ImportWorkflowModal } from './import_workflow_modal';
